+++ client/src/app/tabs/dmn/__tests__/DmnEditorSpec.js
@@ -303,7 +303,8 @@ describe('<DmnEditor>', function() {
           return [];
         },
         onError: onErrorSpy,
-        onAction: noop
+        onAction: noop,
+        emit: noop
       };
 
       // then
@@ -2273,8 +2274,8 @@ describe('<DmnEditor>', function() {
     beforeEach(function() {
       emittedEvents = [];
 
-      recordActions = (action, options) => {
-        emittedEvents.push(options);
+      recordActions = (type, payload) => {
+        emittedEvents.push({ type, payload });
       };
     });
 
@@ -2282,7 +2283,7 @@ describe('<DmnEditor>', function() {
 
       // when
       await renderEditor(diagramXML, {
-        onAction: recordActions
+        emit: recordActions
       });
 
       // then
@@ -2303,7 +2304,7 @@ describe('<DmnEditor>', function() {
       const {
         instance
       } = await renderEditor(diagramXML, {
-        onAction: recordActions
+        emit: recordActions
       });
 
       // then